chore: extract presets.
