+++ brd-ios/breadwallet/src/Redesign/Models/ServerResponse.swift
@@ -17,7 +17,6 @@ struct ServerResponse: Decodable {
     
     struct ServerError: Decodable, FEError {
         var code: String?
-//        var server_mesage: String?
         var serverMesage: String?
         var statusCode: Int {
             return Int(code ?? "") ?? -1
