+++ brd-ios/breadwallet/src/Redesign/Views/TitleValueView.swift
@@ -10,6 +10,16 @@
 
 import UIKit
 
+extension Presets {
+    struct TitleValue {
+        static var horizontal = TitleValueConfiguration(title: .init(font: Fonts.Body.one, textColor: LightColors.Text.two),
+                                                      value: .init(font: Fonts.Body.two, textColor: LightColors.Text.one, textAlignment: .right))
+        
+        static var vertical = TitleValueConfiguration(title: .init(font: Fonts.Body.two, textColor: LightColors.Text.one),
+                                                      value: .init(font: Fonts.Body.two, textColor: LightColors.Text.one))
+    }
+}
+
 struct TitleValueConfiguration: Configurable {
     var title: LabelConfiguration
     var value: LabelConfiguration
